Add ownership transfer protocol
This currently allows moving:
1. The entire uid-space that a group owns, or
2. Splitting a group's owned UID prefix in two (by adding one more bit)
   and moving one of the resulting "halves".

It does not currently allow splitting the UID-space arbitrarily or
merging in an adjacent prefix of UIDs. Those features would be easy to
add if the underlying tree supported them.

The current implementation is pretty dangerous, as described in the
comments in `cluster.rs`. It won't stop you from permanently losing a
prefix of your UIDs on a failed transfer.