Delete reference server
The problem is the reference server was sort of just an Actix interface
atop code that is now copied into `app.rs` for the HSMs. We don't want
the Actix part, and we don't want the reference code to diverge from its
copy in `app.rs`. We'll probably need something like it eventually, and
at that point, we'll rebuild (and/or resurrect) what we actually need,
which may or may not call into `app.rs`.