Move checking auth token from HSM to load balancer
Also rename `rid` to `record_id`. (It started as `uid` but then was
changed to `rid`, which seems more obscure.)

This auth token change is useful for the load balancer to continue to
have visibility into routing the requests, even once the requests are
encrypted. If we collectively decide the HSMs should be the only ones to
see the auth tokens, we can partially undo this.