Update merkle_tree_docgen to run graphviz in Docker
Different dot versions and different fonts can affect the output (though
somehow not always).

dot is now run on all diagrams at the end, since the latency of spawning a
Docker container can be high.

This also updates from Typst 0.5 to 0.9, which requires some minor
changes. The new PDF is different in very trivial ways and in the
bibliography styling.