Use a secret key to randomize record IDs
The purpose is to pseudo-randomly distribute record IDs (even with
adversarially-generated user IDs), so that the Merkle trees stay
balanced. This is similar to how hash tables are randomized.

This requires a new type of secret (different from a tenant auth key) in
Google Cloud Secret Manager and the JSON file equivalent.